+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Bug fixed: Basic Auth (and Oauth BA) asswords are now hashed, this is backward compatible, plaintext passwords will still work
 - OAuth access token expiry can now be set (in seconds) in the `tyk.conf` file using `oauth_token_expire=3600`
 - Proxy now records accurate status codes for upstream requests for better error reporting
+- Added refresh token invalidation API: `DELETE /tyk/oauth/refresh/{key}?api_id={api_id}`
 
 # 1.8.3.2